support max_size attribute
